refactor: defining all responsibilities
