+++ src/controllers/userController.ts
@@ -1,0 +1,8 @@
+import { FastifyRequest } from "fastify"
+
+
+export async function getUser(request: FastifyRequest){
+  await request.jwtVerify()
+
+  return { user: request.user }
+}
