+++ src/routes/auth.ts
@@ -1,8 +1,6 @@
-import {FastifyInstance} from "fastify";
-import {authenticate} from "../plugins/authenticate";
-import { getUser, postUser } from "../controllers/authController";
+import { FastifyInstance } from "fastify";
+import { postUser } from "../controllers/authController";
 
 export async function authRoutes(fastify: FastifyInstance) {
-	fastify.get('/me', {onRequest: [authenticate]} , getUser)
 	fastify.post('/users', async (request) => await postUser(fastify, request))
 }
