+++ src/schemas/authSchema.ts
@@ -1,9 +1,5 @@
 import { z } from "zod";
 
-export const authCreateUserBodySchema = z.object({
-  access_token: z.string()
-})
-
 export const authUserInfoSchema = z.object({
   id: z.string(),
   email: z.string().email(),
