+++ src/schemas/tokenSchema.ts
@@ -1,0 +1,5 @@
+import { z } from "zod";
+
+export const accessTokenBodySchema = z.object({
+  access_token: z.string()
+})
